Move address types to addresses
With the previous changes made, all of the address and error types can
now be moved to the addresses crate.

Move all address types to addresses and re-export in bitcoin.
