+++ addresses/Cargo.toml
@@ -15,16 +15,23 @@ exclude = ["tests", "contrib"]
 
 [features]
 default = ["std"]
-std = ["alloc", "crypto/std", "internals/std", "primitives/std", "taproot_primitives/std"]
-alloc = ["crypto/alloc", "internals/alloc", "primitives/alloc", "taproot_primitives/alloc"]
+std = ["alloc", "base58/std", "bech32/std", "crypto/std", "hashes/std", "internals/std", "network/std", "primitives/std", "serde/std", "taproot_primitives/std"]
+alloc = ["base58/alloc", "bech32/alloc", "crypto/alloc", "hashes/alloc", "internals/alloc", "network/alloc", "primitives/alloc", "serde/alloc", "taproot_primitives/alloc"]
 
 [dependencies]
-crypto = { package = "bitcoin-crypto", path = "../crypto", version = "0.2.0", default-features = false, features = [] }
+base58 = { package = "base58ck", path = "../base58", version = "0.4.0", default-features = false, features = [] }
+bech32 = { version = "0.11.0", default-features = false, features = [] }
+crypto = { package = "bitcoin-crypto", path = "../crypto", version = "0.2.0", default-features = false, features = ["hex"] }
+hashes = { package = "bitcoin_hashes", path = "../hashes", version = "1.0.0", default-features = false, features = ["hex"] }
 internals = { package = "bitcoin-internals", path = "../internals", version = "0.6.0" }
 taproot_primitives = { package = "bitcoin-taproot-primitives", path = "../taproot_primitives", version = "0.1.0", default-features = false, features = [] }
+network = { package = "bitcoin-network-kind", path = "../network", version = "1.0.0", default-features = false, features = []}
 primitives = { package = "bitcoin-primitives", path = "../primitives", version = "0.103.0", default-features = false, features = ["hex"] }
 
+serde = { version = "1.0.195", default-features = false, features = [ "derive" ], optional = true }
+
 [dev-dependencies]
+serde_json = "1.0.68"
 
 [package.metadata.docs.rs]
 all-features = true
